Move some things out into a utils module
