+++ src/lib.rs
@@ -1,8 +1,6 @@
 #![warn(clippy::all)]
 
 use std::{
-    fmt,
-    net::SocketAddr,
     sync::{
         Arc,
         atomic::{AtomicBool, Ordering},
@@ -11,13 +9,7 @@ use std::{
 };
 
 use aide::axum::ApiRouter;
-use axum::{
-    Extension,
-    extract::DefaultBodyLimit,
-    middleware,
-    response::IntoResponse as _,
-    serve::{Listener, ListenerExt as _},
-};
+use axum::{Extension, extract::DefaultBodyLimit, middleware, serve::Listener};
 use diom_authorization::Permissions;
 use diom_core::Monotime;
 use diom_error::Error;
@@ -40,6 +32,10 @@ use crate::{
         metrics::{ConnectionMetrics, ConnectionType, RequestMetrics},
         otel_spans::trace_layer,
     },
+    utils::{
+        BOOTSTRAPPED, axum_tcp_listener, fail_until_bootstrapped, graceful_shutdown_handler,
+        handle_panic,
+    },
     workers::Workers,
 };
 use diom_core::shutdown::{shutting_down_token, start_shut_down};
@@ -50,69 +46,12 @@ pub mod core;
 pub use diom_error as error;
 mod app_state;
 pub mod openapi;
+mod utils;
 pub mod v1;
 mod workers;
 
 pub(crate) use self::app_state::AppState;
-
-async fn graceful_shutdown_handler() {
-    let ctrl_c = async {
-        tokio::signal::ctrl_c()
-            .await
-            .expect("Failed to install Ctrl+C handler");
-    };
-
-    #[cfg(unix)]
-    let sigterm = async {
-        tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate())
-            .expect("Failed to install SIGTERM handler")
-            .recv()
-            .await;
-    };
-
-    #[cfg(not(unix))]
-    let sigterm = std::future::pending::<()>();
-
-    tokio::select! {
-        _ = ctrl_c => {},
-        _ = sigterm => {},
-    }
-
-    tracing::info!("Received shutdown signal. Shutting down gracefully...");
-    start_shut_down();
-}
-
-fn handle_panic(err: Box<dyn std::any::Any + Send + 'static>) -> axum::response::Response {
-    if let Some(err) = err.downcast_ref::<String>() {
-        tracing::error!(?err, "Unhandled panic");
-    } else if let Some(err) = err.downcast_ref::<&'static str>() {
-        tracing::error!(?err, "Unhandled panic");
-    } else {
-        tracing::error!("Unhandled non-string panic");
-    }
-    Error::internal("unhandled internal panic").into_response()
-}
-
-async fn axum_tcp_listener(
-    listener: Option<TcpListener>,
-    listen_address: SocketAddr,
-    conn_metrics: ConnectionMetrics,
-    connection_type: ConnectionType,
-) -> impl Listener<Addr: fmt::Display + fmt::Debug> {
-    let listener = match listener {
-        Some(l) => l,
-        None => TcpListener::bind(listen_address)
-            .await
-            .expect("Error binding to listen_address"),
-    };
-
-    listener.tap_io(move |tcp_stream| {
-        if let Err(err) = tcp_stream.set_nodelay(true) {
-            tracing::warn!("failed to set TCP_NODELAY on incoming connection: {err:#}");
-        }
-        conn_metrics.accepted(connection_type);
-    })
-}
+pub use self::utils::Initialized;
 
 async fn run_interserver(
     cfg: Configuration,
@@ -201,54 +140,6 @@ async fn run_internal(
 /// Run the server with the given configuration
 pub async fn run(cfg: Configuration) {
     run_with_listeners(cfg, None, None, Monotime::initial(), Initialized::new()).await
-}
-
-static BOOTSTRAPPED: AtomicBool = AtomicBool::new(false);
-
-async fn fail_until_bootstrapped(
-    path: axum::extract::MatchedPath,
-    request: axum::extract::Request,
-    next: middleware::Next,
-) -> axum::response::Response {
-    let is_admin_route = path.as_str().starts_with("/api/v1.admin.cluster.");
-    if !(is_admin_route || BOOTSTRAPPED.load(Ordering::Relaxed)) {
-        return Error::not_ready("this node has not yet finished bootstrapping").into_response();
-    }
-
-    next.run(request).await
-}
-
-#[derive(Debug, Clone)]
-pub struct Initialized {
-    inner: Arc<tokio::sync::SetOnce<()>>,
-}
-
-impl Default for Initialized {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Initialized {
-    pub fn new() -> Self {
-        Self {
-            inner: Arc::new(tokio::sync::SetOnce::new()),
-        }
-    }
-
-    fn set(&self) -> Result<(), tokio::sync::SetOnceError<()>> {
-        self.inner.set(())
-    }
-
-    // Wait until initialization is finished or the server shuts down
-    pub async fn wait(self) -> diom_error::Result<()> {
-        let shutting_down_token = shutting_down_token();
-        shutting_down_token
-            .run_until_cancelled(self.inner.wait())
-            .await
-            .copied()
-            .ok_or_else(Error::shutting_down)
-    }
 }
 
 /// Run the server with the given configuration and initial state
